+++ darts/utils/likelihood_models/sklearn.py
@@ -252,6 +252,8 @@ class PoissonLikelihood(SKLearnLikelihood):
 
 
 class QuantileRegression(SKLearnLikelihood):
+    _LIKELIHOOD_TYPE = LikelihoodType.Quantile
+
     def __init__(
         self,
         n_outputs: int,
@@ -291,7 +293,7 @@ class QuantileRegression(SKLearnLikelihood):
         self._median_idx = quantiles.index(0.5)
 
         super().__init__(
-            likelihood_type=LikelihoodType.Quantile,
+            likelihood_type=self._LIKELIHOOD_TYPE,
             parameter_names=quantile_names(self.quantiles),
             n_outputs=n_outputs,
         )
@@ -388,6 +390,70 @@ class QuantileRegression(SKLearnLikelihood):
         return model_output[:, :, :, component_medians].reshape(
             k, n_times, n_components
         )
+
+
+class MultiQuantileRegression(QuantileRegression):
+    _LIKELIHOOD_TYPE = LikelihoodType.MultiQuantile
+
+    def __init__(
+        self,
+        n_outputs: int,
+        quantiles: list[float] | None = None,
+    ):
+        """
+        Multi-output Quantile Regression [1]_.
+
+        This is an extension of :class:`QuantileRegression` for scikit-learn models that natively supports
+        multi-quantile prediction (e.g. CatBoost).
+
+        Parameters
+        ----------
+        n_outputs
+            The number of predicted outputs per model call. ``1`` if ``multi_models=False``, otherwise
+            ``output_chunk_length``.
+        quantiles
+            A list of quantiles. Default: ``[0.01, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.99]``.
+
+        References
+        ----------
+        .. [1] https://en.wikipedia.org/wiki/Quantile_regression
+        """
+        super().__init__(n_outputs=n_outputs, quantiles=quantiles)
+
+        if len(self.quantiles) == 1:
+            raise_log(
+                ValueError(
+                    "'multiquantile' likelihood only supports multiple quantiles. "
+                    "For example `quantiles=[0.05, 0.50, 0.95]`."
+                ),
+                logger,
+            )
+
+    def _estimator_predict(
+        self,
+        model,
+        x: np.ndarray,
+        **kwargs,
+    ) -> np.ndarray:
+        # `x` is of shape (n_series * n_samples, n_regression_features)
+        k = x.shape[0]
+
+        # using CatBoost's native multi-quantile support
+        output: np.ndarray = model.model.predict(x, **kwargs)
+
+        # `output` has two shapes depending on whether `MultiOutputRegressor` is used or not:
+        if output.ndim <= 2:
+            # Case 1: univariate & output_chunk_length == 1, shape is (n_series * n_samples, n_quantiles)
+            # -> (n_series * n_samples, 1, 1, n_quantiles)
+            output = output.reshape(k, 1, 1, -1)
+        else:
+            # Case 2: otherwise, shape is (n_quantiles, n_series * n_samples, n_components * output_chunk_length)
+            # -> (n_quantiles, n_series * n_samples, output_chunk_length, n_components)
+            output = output.reshape(output.shape[0], k, self._n_outputs, -1)
+            # -> (n_series * n_samples, output_chunk_length, n_components, n_quantiles)
+            output = output.transpose(1, 2, 3, 0)
+
+        return output
 
 
 class ClassProbabilityLikelihood(SKLearnLikelihood):
@@ -649,6 +715,8 @@ def _get_likelihood(
         return PoissonLikelihood(n_outputs=n_outputs)
     elif likelihood == LikelihoodType.Quantile.value:
         return QuantileRegression(n_outputs=n_outputs, quantiles=quantiles)
+    elif likelihood == LikelihoodType.MultiQuantile.value:
+        return MultiQuantileRegression(n_outputs=n_outputs, quantiles=quantiles)
     elif likelihood == LikelihoodType.ClassProbability.value:
         return ClassProbabilityLikelihood(n_outputs=n_outputs)
     else: